PERF: Use clang-tidy performance enhancement fixes
SRCDIR=/Users/johnsonhj/src/vxl #My local SRC
BLDDIR=/Users/johnsonhj/src/vxl-bld/ #My local BLD

cd /Users/johnsonhj/src/vxl-bld/
run-clang-tidy.py -extra-arg=-D__clang__ -checks=-*,performance*  -header-filter = .* -fix

+++ v3p/openjpeg2/t1.c
@@ -1171,7 +1171,7 @@ bool t1_encode_cblks(
                                                                         datap[(j * cblk_w) + i] =
                                                                                 fix_mul(
                                                                                 tmp,
-                                                                                8192 * 8192 / ((OPJ_INT32) floor(band->stepsize * 8192))) >> (11 - T1_NMSEDEC_FRACBITS);
+                                                                                8192 * 8192 / ((OPJ_INT32) floorf(band->stepsize * 8192))) >> (11 - T1_NMSEDEC_FRACBITS);
                                                                 }
                                                         }
                                                 }
